reforms directory structure & components
