+++ www/main/app.jsx
@@ -1,0 +1,16 @@
+/* eslint-disable */
+import React from 'react';
+import { StoreProvider } from 'easy-peasy';
+
+import GlobalState from './store/GlobalState';
+import Launchpad from './launchpad';
+
+const App = () => {
+  return (
+    <StoreProvider store={GlobalState}>
+      <Launchpad />
+    </StoreProvider>
+  );
+};
+
+export default App;
