+++ www/main/components/addons/bani-controller/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/BaniController';
