+++ www/main/components/addons/bani-controller/utils/generate-qr-code.js
