+++ www/main/components/addons/bani-controller/utils/index.js
@@ -1,0 +1,2 @@
+export { default as getBaniControllerItems } from './get-bani-controller-items';
+export { default as generateQrCode } from './generate-qr-code';
