+++ www/main/components/addons/ceremonies/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/Ceremonies';
