+++ www/main/components/addons/index.js
@@ -1,0 +1,3 @@
+export { default as SundarGutka } from './sundar-gutka';
+export { default as BaniController } from './bani-controller';
+export { default as Ceremonies } from './ceremonies';
