+++ www/main/components/addons/sundar-gutka/bani-cache.js
@@ -1,0 +1,3 @@
+export default {
+  banis: [],
+};
