+++ www/main/components/addons/sundar-gutka/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/SundarGutka';
