--- www/main/components/bani-controller/overlays/CeremoniesOverlayDialog/index.js
@@ -1,1 +1,0 @@
-export { default } from './components/CeremoniesOverlayDialog';
