--- www/main/components/bani-controller/overlays/DialogWrapper.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const DialogWrapper = ({ onDialogClose, children }) => {
-  return (
-    <div className="backdrop">
-      <button className="close-dialog" onClick={onDialogClose}>
-        <i className="fa fa-times" />
-      </button>
-      {children}
-    </div>
-  );
-};
-
-export default DialogWrapper;
