--- www/main/components/bani-controller/overlays/SundarGutkaOverlayDialog/bani-cache.js
@@ -1,3 +1,0 @@
-export default {
-  baniData: [],
-};
