--- www/main/components/bani-controller/overlays/SundarGutkaOverlayDialog/index.js
@@ -1,1 +1,0 @@
-export { default } from './components/SundarGutkaOverlayDialog';
