--- www/main/components/bani-controller/overlays/SyncOverlayDialog/index.js
@@ -1,1 +1,0 @@
-export { default } from './components/SyncOverlayDialog';
