--- www/main/components/bani-controller/overlays/SyncOverlayDialog/utils/index.js
@@ -1,2 +1,0 @@
-export { default as getSyncItems } from './get-sync-items';
-export { default as generateQrCode } from './generate-qr-code';
