--- www/main/components/bani-controller/overlays/index.js
@@ -1,3 +1,0 @@
-export { default as SundarGutkaOverlayDialog } from './SundarGutkaOverlayDialog';
-export { default as SyncOverlayDialog } from './SyncOverlayDialog';
-export { default as CeremoniesOverlayDialog } from './CeremoniesOverlayDialog';
