--- www/main/components/dashboard/index.js
@@ -1,1 +1,0 @@
-export { default } from './Dashboard';
