+++ www/main/components/viewer/Viewer.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Viewer = () => {
+  return <webview id="main-viewer" src="viewer.html" nodeintegration className="base-ui" />;
+};
+
+export default Viewer;
