+++ www/main/constants/index.js
@@ -1,2 +1,4 @@
 /* eslint-disable import/prefer-default-export */
 export const DEFAULT_OVERLAY = 'none';
+export const nitnemBanis = [2, 4, 6, 9, 10, 20, 21, 23];
+export const popularBanis = [90, 30, 31, 22];
