+++ www/main/sttm-ui/index.js
@@ -1,2 +1,3 @@
 export { default as Switch } from './switch';
 export { default as Tile } from './tile';
+export { default as Overlay } from './overlay';
